+++ MotionProfiling_Project/src/main/java/frc/robot/motionprofile/SplitTrajectories.java
@@ -1,6 +1,7 @@
-package frc.robot;
+package frc.robot.motionprofile;
 
-import frc.robot.PathCreater.Path;
+import frc.robot.RobotConstants;
+import frc.robot.motionprofile.Path;
 import jaci.pathfinder.Trajectory;
 import jaci.pathfinder.modifiers.TankModifier;
